# Document state storage refactoring