… BenchmarkDotNet.Artifacts/results/IdeCoreBenchmarks.ProjectOperationBenchmarks.IterateDocuments-report-github.md
… BenchmarkDotNet.Artifacts/results/IdeCoreBenchmarks.ProjectOperationBenchmarks.IterateDocuments-report.csv
… BenchmarkDotNet.Artifacts/results/IdeCoreBenchmarks.ProjectOperationBenchmarks.IterateDocuments-report.html
… src/Workspaces/Core/Portable/Workspace/Solution/ProjectState.cs
… src/Workspaces/Core/Portable/Workspace/Solution/Solution.cs
… src/Workspaces/Core/Portable/Workspace/Solution/SolutionState.CompilationTracker.State.cs
… src/Workspaces/Core/Portable/Workspace/Solution/SolutionState.CompilationTracker.cs
… src/Workspaces/Core/Portable/Workspace/Solution/SolutionState.cs
… src/Workspaces/Core/Portable/Workspace/Workspace.cs
… src/Workspaces/CoreTest/SolutionTests/SolutionTests.cs
… src/Workspaces/Remote/ServiceHub/Host/RemoteWorkspace.SolutionCreator.cs
… src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Core/Utilities/CompilerUtilities/ImmutableHashMapExtensions.cs