@tmat 2021-02-25
> the WithDocumentText benchmarks looks like they increased more than the error, is that expected / problematic

The difference is not significant. Modifying operations are much less important than read operations.