@tmat 2021-03-01
Oops, I have accidentally merged the commits of this PR in a follow-up PR: https://github.com/dotnet/roslyn/pull/51526

@dibarbet I'll send another PR that implements your suggestions.